Better handling of PRC/RFS regsiter map
Added chassis-specific code in place of the hand editing of the json
file I had been doing. Certain registers with bit fields will be
recognized by name and have the bit fields added when the json is
loaded. Registers that map to peripheral chips and SPI connections will
be excluded from polling. Added GET_... PVs to go with ai and mbbi
records so they can be requested outside the polling loop.