From testing and advice on EPICS tech talk, tweaked asyn config for performance
We had been seeing writes skipped when different values were written to
the same PV quickly. Another detail from the tech talk discussion is
that this is expected behavior when writing to a PV multiple times while
it is still processing. The most recent should "win". Users can also
avoid this by monitoring the .PACT field, or doing a caget to wait for
processing to finish, though speed would be the trade off.