Startup script change for waveforms and other extra parts.
Previously the PRCextra code had a startup command called
scllrfPRCConfigure because I hadn't figured out how to have some chassis
types use a subclass of generated code and others not. Now there's a new
macro called "EXTRA" that's used to construct a more sensible name for
the configure command, scllrfPRCextraConfigure, which doesn't conflict
with the command name in the generated code.